experimental test for SettingsQR Generator integration
